SEQNG-155 Move StepTableContainer to its own class #time 1h
